Add CSRF token to the shipment form
The omission of the CSRF token was preventing the form from being
validated by WTForms. This was not caught by the unit tests because CSRF
protection is disabled in the testing config. This change adds the token
to the form and also implements functional testing to catch cases like
these where the Flask test client may not behave like a real browser.