Send to login page when portal api token expires
When the portal token expires methods like get_scientist_visits() will
raise an exception. This was causing an error page to appear to users.
This change makes flask-login check the validity of the api token on
each request. If the token has expired flask-login will redirect the
user to the login page.

We also add a url prefix of /ship and disable the functional tests if a
display is not available.