Show experiments outside of a 1 day window
The get_scientist_visits() method of the portal API will, by default, only
return EPNs within the next 24 hours. We want users to be able to select
experiments weeks into the future. This change enables that by
explicitly setting the start_time and end_time of the window when
calling get_scientist_visits().