Don't render navigation bar on shipment slip page
The shipment slip page is intended for printing and so does not require
the navigation bar. Its presence was also causing the page to print on
two pages. This change removes the navigation bar. We also shrink the
heading font sizes to reduce the chance the slip will span two pages.